Cambio int->float

+++ calculo.c
@@ -5,19 +5,19 @@
 #include "calculo.h"
 
 
-bool comparar_puntos(int x1, int y1, int x2, int y2, int tolerancia){
+bool comparar_puntos(float x1, float y1, float x2, float y2, float tolerancia){
 
-    int diferencia_x = abs(x1 - x2);
-    int diferencia_y = abs(y1 - y2);
+    float diferencia_x = abs(x1 - x2);
+    float diferencia_y = abs(y1 - y2);
 
     if (diferencia_x <= tolerancia && diferencia_y <= tolerancia) return true;
 
     return false;
 }
 
-float distancia_puntos(int x1, int y1, int x2, int y2) {
-    int dx = x2 - x1;
-    int dy = y2 - y1;
+float distancia_puntos(float x1, float y1, float x2, float y2) {
+    float dx = x2 - x1;
+    float dy = y2 - y1;
     return sqrt(dx * dx + dy * dy);
 }                                                       //Quizas reemplazable con:
 
@@ -25,7 +25,7 @@ float norma(float x, float y) {
     return sqrt(x * x + y * y);
 }   
 
-float distancia_punto_a_recta(int xp1, int yp1, int xp2, int yp2, int x_punto, int y_punto){
+float distancia_punto_a_recta(float xp1, float yp1, float xp2, float yp2, float x_punto, float y_punto){
     
     float numerador = abs((yp2 - yp1) * x_punto - (xp2 - xp1) * y_punto + xp2 * yp1 - yp2 * xp1);
     float denominador = sqrt((yp2 - yp1) * (yp2 - yp1) + (xp2 - xp1) * (xp2 - xp1));
@@ -41,7 +41,7 @@ float Bj(float bj, float sumatoria_bj) {
     return (bj + sumatoria_bj);
 }
 
-float bj(float m, float dt, float b, float g, int pos1, int pos2) {
+float bj(float m, float dt, float b, float g, float pos1, float pos2) {
     return (((m/(dt*dt))*((2*pos1) - pos2))+((b/dt)*pos1) + (m*g));
 }
